Reference

Basics

Nash.random_2players_game — Function

random_2players_game return random payoff matrices for 2 players with given number of actions

Input parameters

dist - distribution from which payoffs are sampled

p1_size - number of actions of first player

p2_size - number of actions of second player

Nash.random_nplayers_game — Function

random_nplayers_game return random payoffs for n players with given number of actions

Input parameters

dist - distribution from which payoffs are sampled

size - vector or tuple of number of players' actions

Nash.outer — Function

outer takes vector of vectors and return outer product tensor

Input parameters

vecs - array of vectors to use in outer product

Nash.get_payoff — Function

get_payoff produce payoff profile for given game and actions probabilities array

Input parameters

game - array of vectors to use in outer product

s - vector of actions probabilities vector